requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
contourpy==1.1.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
distro==1.9.0
Django==4.2.17
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-redis==5.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.15.2
exceptiongroup==1.2.2
fonttools==4.55.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_resources==6.4.5
incremental==24.7.2
jiter==0.8.2
kiwisolver==1.4.7
matplotlib==3.7.5
msgpack==1.1.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
Pillow==9.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyOpenSSL==25.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2022.2.1
redis==5.2.1
scipy==1.10.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
xlrd==1.2.0
zipp==3.20.2
zope.interface==7.2
